backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
numpy
